+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
-### v1.0.0-rc.2
+### v1.0.0-rc.2 (09.09.25)
+
+- Fixed a bug where DOM elements added via actions would not get diffs injected into them. 
 
 - Renamed the `:force` attribute in `m:Tamnoon.DOM.Actions.ToggleAttribute` to `:force_to`.